+++ changelog.md
@@ -5,6 +5,7 @@
 - **New:**
   - Add [#COMMENT](docs/commands.md#miscellaneous) does nothing, used to create quick comments in scripts
   - Add [#NOOP](docs/commands.md#miscellaneous) does nothing, will expand arguments and execute any functions but return nothing to the screen
+  - Add [#UNACTION](docs/commands.md#triggers) same as [#UNTRIGGER](docs/commands.md#triggers)
   - Add [%clip(string)](docs/functions.md#miscellaneous) return or set text on clipboard
   - [Preferences](docs/preferences.md):
     - Add [Ignore Eval Undefined](docs/preferences.md#scripting) Will cause undefined results to be ignored if enabled, else it will be converted to a string and the word undefined displayed
@@ -14,6 +15,7 @@
     - Add [Block Comment String](docs/preferences.md#scripting--special-characters) The 1 or 2 character string for block comments, ending block comment is string reversed
     - Add [Block Inline Comment](docs/preferences.md#scripting--special-characters) disable block comment support    
   - [Triggers](docs/profiles.md#triggers)
+    - Add `Loop Expression` trigger type
     - Add `ReParse` and `ReParse Pattern` sub type triggers
     - Add `Manual` sub type, can only be fired using [#SET](docs/commands.md#triggers) command
     - Add `Wait` sub type trigger
@@ -33,6 +35,8 @@
 - **Changed:**
   - [Triggers](docs/profiles.md#triggers), [Buttons](docs/profiles.md#buttons), [Macros](docs/profiles.md#macros), [Context](docs/profiles.md#context), and [Aliases](docs/profiles.md#aliases) values will now be ignored if empty, thus not creating blank lines, you can use a value of ${cr} to send a blank line
   - [#ADD](docs/commands.md#miscellaneous) will default to 0 if variable is not defined and define variable
+  - Update electron 16.0.7 to 16.0.8
+  - Update @electron/remote 2.0.1 to 2.0.4
 
 ## v0.17.6 2022-01-24